Use folder-remote icon
stock_shared-by-me is deprecated and has a fallback icon which looks ugly.
share-logo.png is a scaled version of that icon. Both are now dropped in favour
of folder-remote which is provided for by the theme.

fix.pl is dropped because we no longer need to prepend the path of the icon,
and share-dialog.glade.in is now share-dialog.glade, since we no longer need to
modify it for installation.