Set the title of the sharing options window
The title of the standalone sharing options window is set to Nautilus's
default, which is "File Manager". This is not translatable because the string
is not present in our gettext domain. This patch sets the title to "Folder
Sharing" manually which is more appropriate than "File Manager" anyway. This
allows the string to be translated